Move struct/enum in parser::types:: to ast::types::
