+++ espr/src/lib.rs
@@ -6,5 +6,6 @@
 
 #![deny(broken_intra_doc_links)]
 
+pub mod ast;
 pub mod parser;
 pub mod semantics;
