+++ espr/src/parser/entity/entity.rs
@@ -118,6 +118,7 @@ pub fn entity_decl(input: &str) -> ParseResult<Entity> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::ast::types::*;
     use nom::Finish;
 
     #[test]
