+++ espr/src/parser/types/concrete.rs
@@ -3,32 +3,6 @@ use super::{
     *,
 };
 
-/// Output of [concrete_types]
-#[derive(Debug, Clone, PartialEq)]
-pub enum ConcreteType {
-    Simple(SimpleType),
-    Reference(String),
-    Set {
-        bound: Option<Bound>,
-        base: Box<ConcreteType>,
-    },
-    Bag {
-        bound: Option<Bound>,
-        base: Box<ConcreteType>,
-    },
-    List {
-        unique: bool,
-        bound: Option<Bound>,
-        base: Box<ConcreteType>,
-    },
-    Array {
-        unique: bool,
-        optional: bool,
-        bound: Bound,
-        base: Box<ConcreteType>,
-    },
-}
-
 /// 193 concrete_types = [aggregation_types] | [simple_types] | [type_ref].
 pub fn concrete_types(input: &str) -> ParseResult<ConcreteType> {
     alt((
@@ -100,12 +74,6 @@ pub fn set_type(input: &str) -> ParseResult<ConcreteType> {
             base: Box::new(base),
         })
         .parse(input)
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct Bound {
-    pub lower: Expression,
-    pub upper: Expression,
 }
 
 /// 183 bound_1 = [numeric_expression] .
